requirements.txt
setuptools~=60.2.0
boto3~=1.34.107
pandas~=2.2.2
scikit-learn~=1.4.2
numpy~=1.26.4
PyYAML~=6.0.1
click~=8.1.7
uvicorn~=0.29.0
fastapi~=0.111.0
requests~=2.31.0
pydantic~=2.7.1